+++ packages/two_percent/docs/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Unreleased
 
+- Make two percent authentication safe by default [#331](https://github.com/powerhome/power-tools/pull/331)
+- Add logger and logger config override [#331](https://github.com/powerhome/power-tools/pull/331)
+
 # [0.4.0] - 2025-05-15
 
 - Add Bulk operations support [#329](https://github.com/powerhome/power-tools/pull/329)